# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Restaurant operations platform API",
    "description": "Orders at the table, tickets live in the kitchen, a bill at the end.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/visits/{id}": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "One table's whole meal.",
        "description": "# Errors\n\nReturns `404` if there is no such visit in this restaurant, which is also\nthe answer for another restaurant's visit so the two cannot be told apart,\n`401` if nobody is signed in, and `403` if the caller is not a waiter.",
        "operationId": "visit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The visit to read.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The visit, its rounds, and its bill. Waiters only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitResponse"
                }
              }
            }
          },
          "401": {
            "description": "Nobody is signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not a waiter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such visit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/visits/{id}/close": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Ends the meal: closes the bill, then closes the visit.",
        "description": "In that order and in one transaction. `close_visit` refuses while a bill on\nthe visit is still open, so the order is not a preference; and one\ntransaction is what stops a bill being numbered and totalled while its table\nstays occupied for ever because the second write failed.\n\n# Errors\n\nReturns `409 bill_has_unserved_lines` if a dish has not reached the table,\n`409 bill_already_closed` if somebody closed it first, `409\nbill_has_no_lines` if nothing was ordered, `404` if there is no such visit or\nit has no bill, `401` if nobody is signed in, and `403` if the caller is not\na waiter.",
        "operationId": "close_visit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The visit to close.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The closed bill, with its number and every figure. Waiters only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillDto"
                }
              }
            }
          },
          "401": {
            "description": "Nobody is signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not a waiter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such visit, or it has no bill.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "A dish is still out, or the bill has already closed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BillDto": {
        "type": "object",
        "description": "A bill, open or closed, with every figure it currently carries.\n\nThe currency travels with it rather than being read from the restaurant,\nbecause a closed bill keeps the currency it was charged in even after the\nrestaurant changes. A screen formatting a bill must use these two and not\nthe restaurant's current pair.",
        "required": [
          "id",
          "status",
          "currencyCode",
          "currencyDecimals",
          "subtotal",
          "serviceChargeAmount",
          "taxTotal",
          "total",
          "taxes"
        ],
        "properties": {
          "closedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When it closed, if it has."
          },
          "currencyCode": {
            "type": "string",
            "description": "What it is charged in."
          },
          "currencyDecimals": {
            "type": "integer",
            "format": "int32",
            "description": "How many decimal places that currency uses.",
            "minimum": 0
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Which bill this is."
          },
          "number": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Its number within the restaurant, allocated at close. `null` while open."
          },
          "serviceChargeAmount": {
            "type": "string",
            "description": "What that rate came to. Zero rather than null when there is none."
          },
          "serviceChargePercent": {
            "type": [
              "string",
              "null"
            ],
            "description": "The service charge rate applied, or `null` when the restaurant charges\nnone."
          },
          "status": {
            "type": "string",
            "description": "Whether it is still collecting dishes."
          },
          "subtotal": {
            "type": "string",
            "description": "The sum of its dishes that were not cancelled."
          },
          "taxTotal": {
            "type": "string",
            "description": "The sum of the tax lines."
          },
          "taxes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BillTaxDto"
            },
            "description": "The tax breakdown, empty until it closes."
          },
          "total": {
            "type": "string",
            "description": "Subtotal plus service charge plus taxes, with no residue."
          }
        }
      },
      "BillTaxDto": {
        "type": "object",
        "description": "One tax line copied onto a bill when it closed.",
        "required": [
          "name",
          "ratePercent",
          "amount"
        ],
        "properties": {
          "amount": {
            "type": "string",
            "description": "What that came to, already rounded to the bill's currency."
          },
          "name": {
            "type": "string",
            "description": "What the tax was called at the time."
          },
          "ratePercent": {
            "type": "string",
            "description": "What rate was charged, as an exact decimal string."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The body every failed request returns.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "A stable machine readable code, safe to branch on in the client.\n\nA refused conflict names which conflict it was, such as\n`table_occupied` or `bill_has_unserved_lines`, rather than the blanket\nword `conflict`. The closed list is\n[`ConflictKind`](crate::domain::error::ConflictKind).",
            "example": "not_found"
          },
          "fields": {
            "type": [
              "object",
              "null"
            ],
            "description": "What is wrong with each named field, when the problem is a form rather\nthan the request as a whole.\n\nAbsent from every response that has no field level problem, which is\nmost of them, so nothing that was already reading this body changes.\nEach value is one of a closed set of codes: `already_taken`,\n`too_short`, `too_long`, `invalid_format`, `unknown_country`,\n`not_in_catalogue`, `incorrect`, `required`.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            },
            "example": {
              "email": "already_taken"
            }
          },
          "message": {
            "type": "string",
            "description": "A human readable sentence. Never contains internal detail.\n\nEnglish, and for a log. The web app renders a translation of\n[`Self::error`] instead, because the API is a data API and knows nothing\nabout what language anybody reads.",
            "example": "The requested resource does not exist."
          }
        }
      },
...
      "LineStatusDto": {
        "type": "string",
        "description": "Where one dish has got to, on the wire.",
        "enum": [
          "queued",
          "ready",
          "served",
          "voided"
        ]
      },
...
      "OrderLineDto": {
        "type": "object",
        "description": "One dish on one ticket, as every screen reads it.\n\nThe name and the price are the ones copied onto the line when the round was\nsent, not today's menu. That is what lets a bill from last Tuesday stay\ntrue after a reprice.",
        "required": [
          "id",
          "dishId",
          "dishName",
          "quantity",
          "unitPrice",
          "lineTotal",
          "status"
        ],
        "properties": {
          "dishId": {
            "type": "string",
            "format": "uuid",
            "description": "Which menu item it came from."
          },
          "dishName": {
            "type": "string",
            "description": "What the dish was called when it was ordered."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Which line this is."
          },
          "lineTotal": {
            "type": "string",
            "description": "Quantity times unit price, as an exact decimal string."
          },
          "note": {
            "type": [
              "string",
              "null"
            ],
            "description": "What the guest asked for, such as no onions."
          },
          "quantity": {
            "type": "integer",
            "format": "int32",
            "description": "How many."
          },
          "status": {
            "$ref": "#/components/schemas/LineStatusDto",
            "description": "Where this one dish has got to."
          },
          "unitPrice": {
            "type": "string",
            "description": "What one of them cost, as an exact decimal string."
          }
        }
      },
      "OrderRoundDto": {
        "type": "object",
        "description": "One ticket with its dishes.",
        "required": [
          "id",
          "visitId",
          "sequenceNo",
          "status",
          "sentAt",
          "lines"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Which ticket this is."
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderLineDto"
            },
            "description": "Every dish on it, oldest first."
          },
          "readyAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When the last dish came off the pass."
          },
          "sentAt": {
            "type": "string",
            "format": "date-time",
            "description": "When it reached the kitchen."
          },
          "sequenceNo": {
            "type": "integer",
            "format": "int32",
            "description": "Its number within the visit, starting at one."
          },
          "servedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When the last dish reached the table."
          },
          "status": {
            "$ref": "#/components/schemas/RoundStatusDto",
            "description": "Where the whole ticket has got to, recomputed from its dishes."
          },
          "visitId": {
            "type": "string",
            "format": "uuid",
            "description": "Which visit it belongs to."
          }
        }
      },
...
      "RoundStatusDto": {
        "type": "string",
        "description": "Where a whole ticket has got to, on the wire.",
        "enum": [
          "queued",
          "ready",
          "served",
          "voided"
        ]
      },
...
      "VisitResponse": {
        "type": "object",
        "description": "One table's whole meal, as the waiter's screen reads it.",
        "required": [
          "id",
          "tableId",
          "tableLabel",
          "status",
          "openedBy",
          "openedAt",
          "rounds",
          "serverTime"
        ],
        "properties": {
          "bill": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BillDto",
                "description": "The bill, with its running subtotal while open and every figure once\nclosed. `null` only for a visit that somehow has none."
              }
            ]
          },
          "guestCount": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "How many of them, when the waiter recorded it."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Which visit this is."
          },
          "openedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When they sat down."
          },
          "openedBy": {
            "type": "string",
            "description": "What to call the waiter who seated them."
          },
          "rounds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderRoundDto"
            },
            "description": "Every ticket on the visit, oldest first, with its dishes."
          },
          "serverTime": {
            "type": "string",
            "format": "date-time",
            "description": "What the server's clock reads, at the moment this answer was built. The\nscreen corrects every age it draws by the difference from its own clock."
          },
          "status": {
            "type": "string",
            "description": "Whether they are still there: `open` or `closed`."
          },
          "tableId": {
            "type": "string",
            "format": "uuid",
            "description": "Which table they are at."
          },
          "tableLabel": {
            "type": "string",
            "description": "What the staff call that table."
          }
        }
      }
    }
  },
...
}